refactor(salvageunion-reference): fold combatUtils into the rules layer
combatUtils.ts was a second doorway onto heat and damage math that the
rules layer already owned. Five of its seven exports had no consumer at
all beyond the barrel re-export and their own tests, and two of those
five duplicated live rules functions: canPush overlapped performPush,
shouldTriggerHeatCheck overlapped performHeatCheck.

Delete lib/combatUtils.ts and relocate its two live functions to the
modules that own their rules:

- canActivateAction -> lib/rules/heatCheck.ts, beside clampHeat and
  performPush, so one module owns the whole heat doorway including the
  pre-flight cap gate.
- applySpDamage -> lib/rules/takeDamage.ts, its only caller's own file,
  still exported so ADR-006's "the SP subtraction is shared, never
  reimplemented" contract keeps a single definition.

Both now leave via the 'salvageunion-reference/rules' subpath rather
than the main barrel, per the note at the foot of lib/index.ts: rules
are deliberately not re-exported from the main entrypoint so consumers
that never touch rules (the Discord bot) don't pull them in. ITUN's two
canActivateAction imports are repointed accordingly.

Deleted with no replacement (zero call sites): getHeatGenerated,
applyHeat, shouldTriggerHeatCheck, canPush, nextCondition. Their 24
tests go with them; the 10 tests covering the two survivors move to
heatCheck.test.ts and takeDamage.test.ts. 3753 -> 3729 passing, 0
failures.

Docs updated to match, including two claims that were already stale:
combat-loop.md credited nextCondition for the condition ladder that
ITUN's own cycleCondition actually drives, and dashboard.md prescribed
gating Push on canPush. That gate is preserved as
canActivateAction(currentHeat, 2, heatCap) — identical arithmetic, no
second symbol — with a note that performPush clamps to cap regardless,
so the gate is a UI affordance rather than a safety net.

No game data touched.
