+++ apps/itun/src/components/dashboard/ActionsDeck.tsx
@@ -17,13 +17,12 @@
 import { useState } from 'react'
 import { ActionsDeck as ActionsDeckView } from 'component-lib'
 import type { ActionsDeckView as ActionsDeckViewModel, DeckGroup } from 'component-lib'
-import { canActivateAction } from 'salvageunion-reference'
 
 import { CORE_ROLL_BANDS, describePushOutcome, performCoreRoll } from '../../lib/rules/coreMechanic'
 import type { CoreRollResult } from '../../lib/rules/coreMechanic'
 import { defaultRoll } from '../../lib/rules/heatCheck'
 import { mechMaxHeat } from '../../lib/rules/derivedStats'
-import { resolveChassisRef } from 'salvageunion-reference/rules'
+import { canActivateAction, resolveChassisRef } from 'salvageunion-reference/rules'
 import type { Mech } from '../../lib/schemas/mech'
 import type { Pilot } from '../../lib/schemas/pilot'
 import { useEntityStore } from '../../stores/entityStore'
