+++ packages/salvageunion-reference/etc/salvageunion-reference.api.d.ts
@@ -181,79 +181,6 @@ export declare function getSchemaCatalog(): {
     schemas: EnhancedSchemaMetadata[];
 };
 //# sourceMappingURL=ModelFactory.d.ts.map
-// === lib/combatUtils.d.ts ===
-/**
- * Combat utility functions for Salvage Union game logic.
- *
- * Pure functions with no side effects. No backend dependency.
- * All combat math and rule checks live here so the logic is testable
- * in isolation and reusable across any consumer (ITUN, Discord bot, etc.).
- */
-import type { ItemCondition } from './utilities.js';
-/**
- * Minimal shape needed to read heat trait data from an entity.
- * The traits array is optional — many entities have no traits.
- */
-type EntityWithTraits = {
-    traits?: Array<{
-        type: string;
-        amount?: number | string;
-    }>;
-    [key: string]: unknown;
-};
-/**
- * Return the heat generated by an entity's hot trait.
- *
- * - Fixed heat:    trait `{ type: 'hot', amount: N }` → returns N
- * - Variable heat: trait `{ type: 'hot (x)' }` → returns `'variable'`
- * - No hot trait:  returns `0`
- */
-export declare function getHeatGenerated(entity: EntityWithTraits): number | 'variable';
-/**
- * Apply heat to a mech, clamping at the heat cap.
- * Heat cannot exceed heatCap.
- */
-export declare function applyHeat(currentHeat: number, heatGenerated: number, heatCap: number): number;
-/**
- * Check whether an action can be activated given current heat and cost.
- * An action is blocked if using it would push heat above the heat cap.
- */
-export declare function canActivateAction(currentHeat: number, heatCost: number, heatCap: number): boolean;
-/**
- * Determine whether a Heat Check is triggered.
- * Per rules p.234, a Heat Check triggers when:
- *  - The mech's heat reaches or exceeds its heat capacity after gaining heat, OR
- *  - The mech is Pushed (Push always brings heat to capacity, so this is covered
- *    by the same condition).
- *
- * Normal heat gain that stays below capacity does NOT trigger a Heat Check.
- */
-export declare function shouldTriggerHeatCheck(currentHeat: number, heatGained: number, heatCapacity: number): boolean;
-/**
- * Check whether the mech can push.
- * Pushing adds 2 heat. Cannot push if the +2 would exceed the heat cap.
- */
-export declare function canPush(currentHeat: number, heatCap: number): boolean;
-/**
- * Return the next condition in the damage progression.
- * intact → damaged → destroyed → destroyed (stays at destroyed)
- */
-export declare function nextCondition(current: ItemCondition): ItemCondition;
-/**
- * Apply SP damage to a mech.
- *
- * Per Salvage Union rules:
- * - newSp = max(0, currentSp - damage)
- * - hpDamage = floor(damage / 2)
- *
- * Returns the new SP value and the HP damage that flows through to the pilot.
- */
-export declare function applySpDamage(currentSp: number, damage: number): {
-    newSp: number;
-    hpDamage: number;
-};
-export {};
-//# sourceMappingURL=combatUtils.d.ts.map
 // === lib/contentBlockHelpers.d.ts ===
 /**
  * Helper functions for working with content blocks in entity data
@@ -669,7 +596,6 @@ export { nameToSlug, getEntitySlug, findEntityBySlug } from './slug.js';
 export { replaceChassisPlaceholder, parseContentBlockString, resolveDataValueForTechLevel, } from './contentBlockHelpers.js';
 export { search, searchIn, getSuggestions, invalidateSearchIndex, type SearchOptions, type SearchResult, } from './search.js';
 export { resolveChoiceView, type ChoiceSelections, type ChoicePrompt, type ResolvedChoiceView, } from './resolveChoiceView.js';
-export { getHeatGenerated, applyHeat, canActivateAction, shouldTriggerHeatCheck, canPush, nextCondition, applySpDamage, } from './combatUtils.js';
 import { type SearchOptions, type SearchResult } from './search.js';
 export type * from './types/index.js';
 export type { SchemaToEntityMap, EntitySchemaName };
@@ -1637,6 +1563,11 @@ import type { HeatCheckEffect, PushResult, ReactorOverloadOutcome, Roll } from '
  */
 export declare function clampHeat(heat: number, cap: number): number;
 /**
+ * Check whether an action can be activated given current heat and cost.
+ * An action is blocked if using it would push heat above the heat cap.
+ */
+export declare function canActivateAction(currentHeat: number, heatCost: number, heatCap: number): boolean;
+/**
  * Maps a Reactor Overload Table d20 roll to its outcome band (p.234-235).
  *   1     → meltdown
  *   2-5   → system-destroyed
@@ -1711,10 +1642,10 @@ export { isWeaponSystem } from './crawlerSystems.js';
 export { resolveCatalogChoiceEntities, isSchemaOnlyCatalogChoice, } from './choiceCatalog.js';
 export { matchesRef, resolveChassisRef, resolveSystemRef, resolveModuleRef, resolveInstalledRef, refDisplayName, } from './resolveRefs.js';
 export { pilotDetailWarnings, mechDetailWarnings, crawlerDetailWarnings } from './detailWarnings.js';
-export { clampHeat, reactorOverloadOutcome, performHeatCheck, performPush } from './heatCheck.js';
+export { clampHeat, canActivateAction, reactorOverloadOutcome, performHeatCheck, performPush, } from './heatCheck.js';
 export { CORE_ROLL_BANDS, coreRollBand, performCoreRoll, describePushOutcome, } from './coreMechanic.js';
 export type { CoreRollBand, CoreRollBandInfo, CoreRollResult } from './coreMechanic.js';
-export { mechEffectiveDamage, applyMechDamage, criticalDamageOutcome, performCriticalDamage, pilotEffectiveDamage, applyPilotDamage, criticalInjuryOutcome, performCriticalInjury, } from './takeDamage.js';
+export { applySpDamage, mechEffectiveDamage, applyMechDamage, criticalDamageOutcome, performCriticalDamage, pilotEffectiveDamage, applyPilotDamage, criticalInjuryOutcome, performCriticalInjury, } from './takeDamage.js';
 export type { DamageKind, MechDamageInput, MechDamageEffect, PilotDamageInput, PilotDamageEffect, CriticalDamageEffect, CriticalInjuryEffect, } from './takeDamage.js';
 export { PILOT_BASE_HP, PILOT_BASE_AP, PILOT_BASE_INVENTORY_SLOTS, injuryMaxHpPenalty, pilotMaxHP, pilotMaxAP, isPilotDead, clampPilotCurrentStats, installedStatBonus, mechMaxSP, mechMaxEP, mechMaxHeat, mechMaxCargo, clampMechCurrentStats, unifiedMechConditions, crawlerMaxSP, crawlerMaxSPParts, clampCrawlerCurrentStats, } from './derivedStats.js';
 export type { ChassisStats, CrawlerMaxSPParts } from './derivedStats.js';
@@ -2193,8 +2124,8 @@ export declare function evaluateSoftWarnings(before: PilotSnapshot | MechSnapsho
  * The band data mirrors the "Critical Damage" / "Critical Injury" tables in
  * salvageunion-reference data/roll-tables.json, encoded here as pure band
  * functions the same way heatCheck.ts encodes the Reactor Overload bands.
- * The SP subtraction itself is the shared `applySpDamage` from
- * salvageunion-reference/lib/combatUtils.ts (ADR-006 — never reimplemented).
+ * The SP subtraction itself is the shared `applySpDamage` defined below — the
+ * one doorway for SP subtraction (ADR-006 — never reimplemented).
  *
  * This module is PURE: no React, no store, no real randomness — the d20 is
  * injected via `Roll`. Per ADR-007 the functions return deterministic
@@ -2205,6 +2136,19 @@ export declare function evaluateSoftWarnings(before: PilotSnapshot | MechSnapsho
 import type { CriticalDamageOutcome, CriticalDamageResult, CriticalInjuryOutcome, CriticalInjuryResult, Roll } from './types.js';
 /** What the incoming weapon lists in its profile (drives the p.240/241 conversions). */
 export type DamageKind = 'sp' | 'hp';
+/**
+ * Apply SP damage to a mech.
+ *
+ * Per Salvage Union rules:
+ * - newSp = max(0, currentSp - damage)
+ * - hpDamage = floor(damage / 2)
+ *
+ * Returns the new SP value and the HP damage that flows through to the pilot.
+ */
+export declare function applySpDamage(currentSp: number, damage: number): {
+    newSp: number;
+    hpDamage: number;
+};
 /**
  * The SP damage a mech actually takes: HP-listed weapons deal half damage to
  * mechs (floored), then Vulnerable doubles what the mech takes. Non-positive
