--- packages/salvageunion-reference/lib/combatUtils.ts
@@ -1,110 +1,0 @@
-/**
- * Combat utility functions for Salvage Union game logic.
- *
- * Pure functions with no side effects. No backend dependency.
- * All combat math and rule checks live here so the logic is testable
- * in isolation and reusable across any consumer (ITUN, Discord bot, etc.).
- */
-
-import type { ItemCondition } from './utilities.js'
-
-/**
- * Minimal shape needed to read heat trait data from an entity.
- * The traits array is optional — many entities have no traits.
- */
-type EntityWithTraits = {
-  traits?: Array<{ type: string; amount?: number | string }>
-  [key: string]: unknown
-}
-
-/**
- * Return the heat generated by an entity's hot trait.
- *
- * - Fixed heat:    trait `{ type: 'hot', amount: N }` → returns N
- * - Variable heat: trait `{ type: 'hot (x)' }` → returns `'variable'`
- * - No hot trait:  returns `0`
- */
-export function getHeatGenerated(entity: EntityWithTraits): number | 'variable' {
-  const traits = entity.traits
-  if (!traits || traits.length === 0) return 0
-
-  const hotTrait = traits.find((t) => t.type === 'hot' || t.type === 'hot (x)')
-  if (!hotTrait) return 0
-
-  if (hotTrait.type === 'hot (x)') return 'variable'
-
-  const amount = hotTrait.amount
-  if (typeof amount === 'number') return amount
-
-  return 0
-}
-
-/**
- * Apply heat to a mech, clamping at the heat cap.
- * Heat cannot exceed heatCap.
- */
-export function applyHeat(currentHeat: number, heatGenerated: number, heatCap: number): number {
-  return Math.min(currentHeat + heatGenerated, heatCap)
-}
-
-/**
- * Check whether an action can be activated given current heat and cost.
- * An action is blocked if using it would push heat above the heat cap.
- */
-export function canActivateAction(currentHeat: number, heatCost: number, heatCap: number): boolean {
-  return currentHeat + heatCost <= heatCap
-}
-
-/**
- * Determine whether a Heat Check is triggered.
- * Per rules p.234, a Heat Check triggers when:
- *  - The mech's heat reaches or exceeds its heat capacity after gaining heat, OR
- *  - The mech is Pushed (Push always brings heat to capacity, so this is covered
- *    by the same condition).
- *
- * Normal heat gain that stays below capacity does NOT trigger a Heat Check.
- */
-export function shouldTriggerHeatCheck(
-  currentHeat: number,
-  heatGained: number,
-  heatCapacity: number
-): boolean {
-  if (heatGained <= 0) return false
-  return currentHeat + heatGained >= heatCapacity
-}
-
-/**
- * Check whether the mech can push.
- * Pushing adds 2 heat. Cannot push if the +2 would exceed the heat cap.
- */
-export function canPush(currentHeat: number, heatCap: number): boolean {
-  return currentHeat + 2 <= heatCap
-}
-
-/**
- * Return the next condition in the damage progression.
- * intact → damaged → destroyed → destroyed (stays at destroyed)
- */
-export function nextCondition(current: ItemCondition): ItemCondition {
-  if (current === 'intact') return 'damaged'
-  if (current === 'damaged') return 'destroyed'
-  return 'destroyed'
-}
-
-/**
- * Apply SP damage to a mech.
- *
- * Per Salvage Union rules:
- * - newSp = max(0, currentSp - damage)
- * - hpDamage = floor(damage / 2)
- *
- * Returns the new SP value and the HP damage that flows through to the pilot.
- */
-export function applySpDamage(
-  currentSp: number,
-  damage: number
-): { newSp: number; hpDamage: number } {
-  const newSp = Math.max(0, currentSp - damage)
-  const hpDamage = Math.floor(damage / 2)
-  return { newSp, hpDamage }
-}
