+++ packages/salvageunion-reference/lib/index.ts
@@ -76,17 +76,6 @@ export {
   type ResolvedChoiceView,
 } from './resolveChoiceView.js'
 
-// Export combat utility functions (pure game logic)
-export {
-  getHeatGenerated,
-  applyHeat,
-  canActivateAction,
-  shouldTriggerHeatCheck,
-  canPush,
-  nextCondition,
-  applySpDamage,
-} from './combatUtils.js'
-
 // Import search functions for use in class methods
 import {
   search as searchFn,
